+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
+## 1.0.1
+
+- Refactor
+
 ## 1.0.0
-- Initial release: extract grid sharing from EnhancedUpgradeMenu
-- Copy Grid / Paste Code buttons on gear details
+
+### Added
+
+- Initial release: grid sharing extracted from Enhanced Upgrade Menu
+- Copy Grid / Paste Code buttons on the gear action bar (via SparrohUILib)
 - Compact Base64 build codes for equipped upgrade layouts